+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Bugfixes:
 
   - fix to handle multibyte UTF-8 URI decoding
   - warn users if they try to close an active terminal
+  - warn users when their websocket connection is terminated